Create the about.html file
Built the header section of the about.html and styled it by making it responsive. Renamed the styling of the cta from .write-up .cta to .cta in the general style and also in the max-width of 360px. added the image needed for the contact page hero-section. added white-background color to the nav-link.