Create the Contact.html file
created the contact.html and added the image for smaller screen version on the 938px of the contact.html. added the images for the background of the contact and about pages into the image file and attached it to the 938px screen in the css. also formatted the css file so that it looks good to the eyes.